+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
-## 1.4.0 - 2025-02-14
+## 1.4.0 - 2025-02-15
 
 * Added
   * Board: JPEG/PNG preview of TOP/BOTTOM side
   * PnP editor - paginated (200 items/page)
   * PnP editor - improved component matching: CC1206x100n -> C1206 (before: 1206)
+  * PnP editor - filtered view (All / Configured / Removed)
 * Changed
 * Deprecated
 * Removed